Make specs match module names
